[UPDATE] 更新shiro的一些配置 增加角色控制和权限 修复了一些bug

+++ mkshop-manager/mkshop-manager-interface/src/main/java/com/nexus/manager/service/AddressService.java
@@ -16,10 +16,10 @@ import java.util.List;
 public interface AddressService {
 
     //返回用户收货地址
-    public List<TbAddress> getAllAddress(Integer userId);
+    public List<TbAddress> getAllAddress(long userId);
     //更新某用户收货地址
-    public Integer updateAddress(Integer addressId,Integer userId);
+    public Integer updateAddress(long addressId,long userId);
     //删除收货地址
-    public Integer deleteAddress(Integer addressId,Integer userId);
+    public Integer deleteAddress(long addressId,long userId);
 
 }
